refactor(tests): split remaining large test files by endpoint
GuildMembersTests (889 lines) → GuildInvitesTests, LeaveGuildTests, TransferOwnershipTests, RemoveMemberTests, UpdateMemberRoleTests, SendMessageEdgeCaseTests
ChannelMessagesTests (799 lines) → MessageAttachmentsTests, EditMessageTests, DeleteMessageTests, JoinVoiceChannelTests, ReorderChannelsTests
ConversationMessageEndpointsTests (759 lines) → SendConversationMessageTests, GetConversationMessagesTests, EditConversationMessageTests, DeleteConversationMessageTests, ConversationMessageAttachmentsTests
UsersEndpointsTests (597 lines) → UserProfileTests, UserStatusTests

No test logic changes — just split and move.
