+++ tests/Harmonie.API.IntegrationTests/GetConversationMessagesTests.cs
@@ -1,0 +1,238 @@
+using System.Net;
+using System.Net.Http.Headers;
+using System.Net.Http.Json;
+using FluentAssertions;
+using Harmonie.Application.Common;
+using Harmonie.Application.Features.Auth.Register;
+using Harmonie.Application.Features.Conversations.GetMessages;
+using Harmonie.Application.Features.Conversations.OpenConversation;
+using Harmonie.Application.Features.Conversations.SendMessage;
+using Harmonie.Infrastructure.Persistence;
+using Microsoft.AspNetCore.Mvc.Testing;
+using Microsoft.Extensions.DependencyInjection;
+using Xunit;
+
+namespace Harmonie.API.IntegrationTests;
+
+public sealed class GetConversationMessagesTests : IClassFixture<WebApplicationFactory<Program>>
+{
+    private readonly WebApplicationFactory<Program> _factory;
+    private readonly HttpClient _client;
+
+    public GetConversationMessagesTests(WebApplicationFactory<Program> factory)
+    {
+        _factory = factory;
+        _client = factory.CreateClient();
+    }
+
+    [Fact]
+    public async Task GetConversationMessages_WhenCallerIsParticipant_ShouldReturnMessagesAscending()
+    {
+        var caller = await RegisterAsync();
+        var target = await RegisterAsync();
+        var conversationId = await OpenConversationAsync(caller.AccessToken, target.UserId);
+
+        await SendConversationMessageAsync(conversationId, "first direct", caller.AccessToken);
+        await Task.Delay(20);
+        await SendConversationMessageAsync(conversationId, "second direct", target.AccessToken);
+
+        var response = await SendAuthorizedGetAsync(
+            $"/api/conversations/{conversationId}/messages",
+            caller.AccessToken);
+
+        response.StatusCode.Should().Be(HttpStatusCode.OK);
+
+        var payload = await response.Content.ReadFromJsonAsync<GetMessagesResponse>();
+        payload.Should().NotBeNull();
+        payload!.ConversationId.Should().Be(conversationId);
+        payload.Items.Select(x => x.Content).Should().Equal("first direct", "second direct");
+    }
+
+    [Fact]
+    public async Task GetConversationMessages_WhenConversationDoesNotExist_ShouldReturnNotFound()
+    {
+        var caller = await RegisterAsync();
+
+        var response = await SendAuthorizedGetAsync(
+            $"/api/conversations/{Guid.NewGuid()}/messages",
+            caller.AccessToken);
+
+        response.StatusCode.Should().Be(HttpStatusCode.NotFound);
+
+        var error = await response.Content.ReadFromJsonAsync<ApplicationError>();
+        error.Should().NotBeNull();
+        error!.Code.Should().Be(ApplicationErrorCodes.Conversation.NotFound);
+    }
+
+    [Fact]
+    public async Task GetConversationMessages_WhenCallerIsNotParticipant_ShouldReturnForbidden()
+    {
+        var participantOne = await RegisterAsync();
+        var participantTwo = await RegisterAsync();
+        var outsider = await RegisterAsync();
+        var conversationId = await OpenConversationAsync(participantOne.AccessToken, participantTwo.UserId);
+
+        var response = await SendAuthorizedGetAsync(
+            $"/api/conversations/{conversationId}/messages",
+            outsider.AccessToken);
+
+        response.StatusCode.Should().Be(HttpStatusCode.Forbidden);
+
+        var error = await response.Content.ReadFromJsonAsync<ApplicationError>();
+        error.Should().NotBeNull();
+        error!.Code.Should().Be(ApplicationErrorCodes.Conversation.AccessDenied);
+    }
+
+    [Fact]
+    public async Task GetConversationMessages_WithCursorPagination_ShouldReturnNextPage()
+    {
+        var caller = await RegisterAsync();
+        var target = await RegisterAsync();
+        var conversationId = await OpenConversationAsync(caller.AccessToken, target.UserId);
+
+        await SendConversationMessageAsync(conversationId, "first page item", caller.AccessToken);
+        await Task.Delay(20);
+        await SendConversationMessageAsync(conversationId, "second page item", target.AccessToken);
+        await Task.Delay(20);
+        await SendConversationMessageAsync(conversationId, "third page item", caller.AccessToken);
+
+        var firstResponse = await SendAuthorizedGetAsync(
+            $"/api/conversations/{conversationId}/messages?limit=2",
+            caller.AccessToken);
+
+        firstResponse.StatusCode.Should().Be(HttpStatusCode.OK);
+
+        var firstPayload = await firstResponse.Content.ReadFromJsonAsync<GetMessagesResponse>();
+        firstPayload.Should().NotBeNull();
+        firstPayload!.Items.Select(x => x.Content).Should().Equal("second page item", "third page item");
+        firstPayload.NextCursor.Should().NotBeNullOrWhiteSpace();
+
+        var secondResponse = await SendAuthorizedGetAsync(
+            $"/api/conversations/{conversationId}/messages?cursor={Uri.EscapeDataString(firstPayload.NextCursor!)}&limit=2",
+            caller.AccessToken);
+
+        secondResponse.StatusCode.Should().Be(HttpStatusCode.OK);
+
+        var secondPayload = await secondResponse.Content.ReadFromJsonAsync<GetMessagesResponse>();
+        secondPayload.Should().NotBeNull();
+        secondPayload!.Items.Select(x => x.Content).Should().Equal("first page item");
+        secondPayload.NextCursor.Should().BeNull();
+    }
+
+    [Fact]
+    public async Task GetConversationMessages_ShouldExcludeSoftDeletedMessages()
+    {
+        var caller = await RegisterAsync();
+        var target = await RegisterAsync();
+        var conversationId = await OpenConversationAsync(caller.AccessToken, target.UserId);
+
+        var visibleMessage = await SendConversationMessageAsync(conversationId, "visible direct", caller.AccessToken);
+        await Task.Delay(20);
+        var deletedMessage = await SendConversationMessageAsync(conversationId, "deleted direct", target.AccessToken);
+
+        await SoftDeleteConversationMessageAsync(deletedMessage.MessageId);
+
+        var response = await SendAuthorizedGetAsync(
+            $"/api/conversations/{conversationId}/messages",
+            caller.AccessToken);
+
+        response.StatusCode.Should().Be(HttpStatusCode.OK);
+
+        var payload = await response.Content.ReadFromJsonAsync<GetMessagesResponse>();
+        payload.Should().NotBeNull();
+        payload!.Items.Should().ContainSingle();
+        payload.Items[0].MessageId.Should().Be(visibleMessage.MessageId);
+        payload.Items[0].Content.Should().Be("visible direct");
+    }
+
+    [Fact]
+    public async Task GetConversationMessages_WithoutAuthentication_ShouldReturnUnauthorized()
+    {
+        var response = await _client.GetAsync(
+            $"/api/conversations/{Guid.NewGuid()}/messages");
+
+        response.StatusCode.Should().Be(HttpStatusCode.Unauthorized);
+    }
+
+    private async Task<string> OpenConversationAsync(string accessToken, string targetUserId)
+    {
+        var response = await SendAuthorizedPostAsync(
+            "/api/conversations",
+            new OpenConversationRequest(targetUserId),
+            accessToken);
+        response.StatusCode.Should().BeOneOf(HttpStatusCode.Created, HttpStatusCode.OK);
+
+        var payload = await response.Content.ReadFromJsonAsync<OpenConversationResponse>();
+        payload.Should().NotBeNull();
+        return payload!.ConversationId;
+    }
+
+    private async Task<RegisterResponse> RegisterAsync()
+    {
+        var request = new RegisterRequest(
+            Email: $"test{Guid.NewGuid():N}@harmonie.chat",
+            Username: $"user{Guid.NewGuid():N}"[..20],
+            Password: "Test123!@#");
+
+        var response = await _client.PostAsJsonAsync("/api/auth/register", request);
+        response.StatusCode.Should().Be(HttpStatusCode.Created);
+
+        var payload = await response.Content.ReadFromJsonAsync<RegisterResponse>();
+        payload.Should().NotBeNull();
+        return payload!;
+    }
+
+    private async Task<SendMessageResponse> SendConversationMessageAsync(
+        string conversationId,
+        string content,
+        string accessToken)
+    {
+        var response = await SendAuthorizedPostAsync(
+            $"/api/conversations/{conversationId}/messages",
+            new SendMessageRequest(content),
+            accessToken);
+        response.StatusCode.Should().Be(HttpStatusCode.Created);
+
+        var payload = await response.Content.ReadFromJsonAsync<SendMessageResponse>();
+        payload.Should().NotBeNull();
+        return payload!;
+    }
+
+    private async Task<HttpResponseMessage> SendAuthorizedPostAsync<TRequest>(
+        string uri,
+        TRequest payload,
+        string accessToken)
+    {
+        using var request = new HttpRequestMessage(HttpMethod.Post, uri)
+        {
+            Content = JsonContent.Create(payload)
+        };
+        request.Headers.Authorization = new AuthenticationHeaderValue("Bearer", accessToken);
+        return await _client.SendAsync(request);
+    }
+
+    private async Task<HttpResponseMessage> SendAuthorizedGetAsync(
+        string uri,
+        string accessToken)
+    {
+        using var request = new HttpRequestMessage(HttpMethod.Get, uri);
+        request.Headers.Authorization = new AuthenticationHeaderValue("Bearer", accessToken);
+        return await _client.SendAsync(request);
+    }
+
+    private async Task SoftDeleteConversationMessageAsync(string messageId)
+    {
+        await using var scope = _factory.Services.CreateAsyncScope();
+        var dbSession = scope.ServiceProvider.GetRequiredService<DbSession>();
+        var connection = await dbSession.GetOpenConnectionAsync();
+        await using var command = connection.CreateCommand();
+        command.CommandText = """
+                              UPDATE messages
+                              SET deleted_at_utc = @DeletedAtUtc
+                              WHERE id = @MessageId
+                              """;
+        command.Parameters.AddWithValue("DeletedAtUtc", DateTime.UtcNow);
+        command.Parameters.AddWithValue("MessageId", Guid.Parse(messageId));
+        await command.ExecuteNonQueryAsync();
+    }
+}
