+++ tests/Harmonie.API.IntegrationTests/SendMessageEdgeCaseTests.cs
@@ -1,0 +1,145 @@
+using System.Net;
+using System.Net.Http.Headers;
+using System.Net.Http.Json;
+using System.Text.Json;
+using System.Text.Json.Serialization;
+using FluentAssertions;
+using Harmonie.Application.Common;
+using Harmonie.Application.Features.Auth.Register;
+using Harmonie.Application.Features.Channels.SendMessage;
+using Harmonie.Application.Features.Guilds.CreateGuild;
+using Harmonie.Application.Features.Guilds.GetGuildChannels;
+using Microsoft.AspNetCore.Mvc.Testing;
+using Xunit;
+
+namespace Harmonie.API.IntegrationTests;
+
+public sealed class SendMessageEdgeCaseTests : IClassFixture<WebApplicationFactory<Program>>
+{
+    private readonly HttpClient _client;
+    private static readonly JsonSerializerOptions _jsonOptions = new()
+    {
+        Converters = { new JsonStringEnumConverter() }
+    };
+
+    public SendMessageEdgeCaseTests(WebApplicationFactory<Program> factory)
+    {
+        _client = factory.CreateClient();
+    }
+
+    [Fact]
+    public async Task SendMessage_WhenChannelIsVoice_ShouldReturnConflict()
+    {
+        var user = await RegisterAsync();
+
+        var createGuildResponse = await SendAuthorizedPostAsync(
+            "/api/guilds",
+            new CreateGuildRequest("Voice Guild"),
+            user.AccessToken);
+        createGuildResponse.StatusCode.Should().Be(HttpStatusCode.Created);
+
+        var createGuildPayload = await createGuildResponse.Content.ReadFromJsonAsync<CreateGuildResponse>();
+        createGuildPayload.Should().NotBeNull();
+
+        var channelsResponse = await SendAuthorizedGetAsync(
+            $"/api/guilds/{createGuildPayload!.GuildId}/channels",
+            user.AccessToken);
+        channelsResponse.StatusCode.Should().Be(HttpStatusCode.OK);
+
+        var channelsPayload = await channelsResponse.Content.ReadFromJsonAsync<GetGuildChannelsResponse>();
+        channelsPayload.Should().NotBeNull();
+
+        var voiceChannel = channelsPayload!.Channels.First(channel => channel.Type == "Voice");
+
+        var sendMessageResponse = await SendAuthorizedPostAsync(
+            $"/api/channels/{voiceChannel.ChannelId}/messages",
+            new SendMessageRequest("Should fail"),
+            user.AccessToken);
+        sendMessageResponse.StatusCode.Should().Be(HttpStatusCode.Conflict);
+
+        var error = await sendMessageResponse.Content.ReadFromJsonAsync<ApplicationError>();
+        error.Should().NotBeNull();
+
+        error!.Code.Should().Be(ApplicationErrorCodes.Channel.NotText);
+    }
+
+    [Fact]
+    public async Task SendMessage_WhenRateLimitExceeded_ShouldReturnTooManyRequests()
+    {
+        var user = await RegisterAsync();
+
+        var createGuildResponse = await SendAuthorizedPostAsync(
+            "/api/guilds",
+            new CreateGuildRequest("Rate Limit Guild"),
+            user.AccessToken);
+        createGuildResponse.StatusCode.Should().Be(HttpStatusCode.Created);
+
+        var createGuildPayload = await createGuildResponse.Content.ReadFromJsonAsync<CreateGuildResponse>();
+        createGuildPayload.Should().NotBeNull();
+
+        var channelsResponse = await SendAuthorizedGetAsync(
+            $"/api/guilds/{createGuildPayload!.GuildId}/channels",
+            user.AccessToken);
+        channelsResponse.StatusCode.Should().Be(HttpStatusCode.OK);
+
+        var channelsPayload = await channelsResponse.Content.ReadFromJsonAsync<GetGuildChannelsResponse>();
+        channelsPayload.Should().NotBeNull();
+
+        var textChannel = channelsPayload!.Channels.First(channel => channel.Type == "Text");
+
+        for (var i = 0; i < 40; i++)
+        {
+            var sendResponse = await SendAuthorizedPostAsync(
+                $"/api/channels/{textChannel.ChannelId}/messages",
+                new SendMessageRequest($"msg-{i}"),
+                user.AccessToken);
+
+            sendResponse.StatusCode.Should().Be(HttpStatusCode.Created);
+        }
+
+        var throttledResponse = await SendAuthorizedPostAsync(
+            $"/api/channels/{textChannel.ChannelId}/messages",
+            new SendMessageRequest("msg-over-limit"),
+            user.AccessToken);
+
+        throttledResponse.StatusCode.Should().Be((HttpStatusCode)429);
+    }
+
+    private async Task<RegisterResponse> RegisterAsync()
+    {
+        var request = new RegisterRequest(
+            Email: $"test{Guid.NewGuid():N}@harmonie.chat",
+            Username: $"user{Guid.NewGuid():N}"[..20],
+            Password: "Test123!@#");
+
+        var response = await _client.PostAsJsonAsync("/api/auth/register", request);
+        response.StatusCode.Should().Be(HttpStatusCode.Created);
+
+        var payload = await response.Content.ReadFromJsonAsync<RegisterResponse>();
+        payload.Should().NotBeNull();
+
+        return payload!;
+    }
+
+    private async Task<HttpResponseMessage> SendAuthorizedPostAsync<TRequest>(
+        string uri,
+        TRequest payload,
+        string accessToken)
+    {
+        using var request = new HttpRequestMessage(HttpMethod.Post, uri)
+        {
+            Content = JsonContent.Create(payload, options: _jsonOptions)
+        };
+        request.Headers.Authorization = new AuthenticationHeaderValue("Bearer", accessToken);
+        return await _client.SendAsync(request);
+    }
+
+    private async Task<HttpResponseMessage> SendAuthorizedGetAsync(
+        string uri,
+        string accessToken)
+    {
+        using var request = new HttpRequestMessage(HttpMethod.Get, uri);
+        request.Headers.Authorization = new AuthenticationHeaderValue("Bearer", accessToken);
+        return await _client.SendAsync(request);
+    }
+}
